+++ docs/change_log.md
@@ -1,6 +1,20 @@
 # Changelog
 All notable changes to this project will be documented in this file. We keep track of changes in this file since v0.1.8. The source code for most releases is available on [GitHub](https://github.com/mikekeith52/scalecast).
 
+## [0.9.9] - 
+### Added
+### Changed
+- warning logs now called after `Forecaster` and `MVForecaster` objects are initiated
+- No more `get_funcs()` method, a bid idea from the start
+### Fixed
+
+
+## [0.9.8] - 2022-05-11
+### Added
+- added the `corr()` and `corr_lags()` methods to the `MVForecaster` object
+### Changed
+### Fixed
+
 ## [0.9.7] - 2022-05-10
 ### Added
 ### Changed